Plate data read in and storage functions
Created two functions. One for searching the raw data file and determinig its type (plate 1 or 2) and finding the beggining of the data and metadata tables and their lenghs. The other for extracting the data and metadata tables into pandas data frames, coonverting them into one column data frames and storing in a dictionary.