Plate data read in  and initial processing
Created two functions. One for searching the raw data file and determinig its type (list A, B, or C) and finding the beggining of the data and metadata tables and their lenghs. The other, for extracting the data and metadata tables into pandas data frames, converting them into one column data frames and storing in a dictionary.