Update fluores_anisotropy.ipynb
Modified the mechanism for removing the middle '0' from the wellnumber in the process_list function.  The for loop was replaced with a pandas .apply() method and a lambda function.

Connverted both read in functions into classmethods and added an if-else statement in the __init__ method to execute the appropriate pre-processing function and assign its output to the class attribute 'self.data_dict'.