Update fluores_anisotropy.ipynb
Fixed the bug in the read in list function which caused an error during reading in data files in which the g-factor was placed after the data table. Now the 'end_of_data' parameter is defined for each if-else statement by filtering the blank indiced list and  finding the first blank index greater than 'skiprows' parameter.