Update fluores_anisotropy.ipynb
The mechanism of generating the well ids was moved to the read_in_envision function because it is now common to both read in functions.
Tuple unpacking is now used to assign the output of read in functions to the data_dict and g factor variables in the read in envision classmethod.
Comments were updated.