Update fluores_anisotropy.ipynb
Created invalidate function to manage the three invalidate functions from the platemapping package.
Created visualise function which call the visualise function from yhe platemapping package and implemented functionaality to label or color the platemap by the 'p' or 's' values of anisotropy.
Added a header names dictionary.
Updated the read in envision function so that the platemap is executed after the data read in and the class is initialised with another parameter - the platemap.
Added a summary of data that is displayed after data read in: g-factor, number of repeats and number of empty cells.