Update plate map.ipynb
Removed the data type dictionary and created a new one in the empty plate function based on the header names dictionary.
Fixed a bug in the plate_map function which caused error if the plate map csv file had a column with all cells 'nan' in the trailing spaces removal mechanism.
Updated the labelwell function by adding option to print the 's' and 'p' channel values in scientific notation.