Update fluores_anisotropy.ipynb
Resolved a bug that cause error when trying to read in data in plate 2 format but with commas instead of blank lines. Added another condition in to the list comprehension that generates the list of blank rows to consider lines with commas only as blank lines. Relpaced the re.findall method with re.fullmatch method to find the g-factor.  Added an if statement to delete the last column if there are 25 or 13 coumns in raw data in the pre process plate function.