Update test_fluores_anisotropy.ipynb
Added two new tests for the calculate_r_i function. One to test for the correct values of calculated anisotropy and intensity and the othr to check if an error is raised when the functoin is called with corrected parameter as 'True' but the background subtraction has not been performed.