Added plot_i_percent function
The ploti_percent functon was added which returns the _plot_i_percent funcotin so that results of i  percent calculation can be displyed after the calculation of i percent.
The _plot_i_percent function now iterates over the the data dictionary instead of list of tuples and it checks for well ids above the threshold itself. The threshold value is now stored as a class variable.