Update fluores_anisotropy.ipynb
The set_fitparams functoin was updated to check if the data frame with fitting paramers exists, otherwise error is raised.
In the calc_data_to_fit function the column names of the fitting data frame were corrected, default lambda value was set as 1, the errors of fitting parameters are calculated and included in the fitting parameters data frame.