Created functions for plotting the anisotropy and itensity aginst protein concentration
The first set of functions plots the anisotropy and intensity as subplots and displays the two fully labelled graphs for each protein-tracer pair side by side for easy comparison. A separate figure with subplots is generated for each repeat.
The second set of functions a single figure anisotropy or intensity for all repeats and saves it in a default or user defined directory.