Update fluores_anisotropy.ipynb
Another parameter was added to the logistic_fit_all and logistic_fit_pair functions to enable the curve fitting with the ‘sigma’ parameter in the SciPy curve_fit function. The _logistic_fit function was modified to extract the standard deviation or standard error data and pass it to the curve_fit function.
The widget to display the quality score of calculated lambda value was added to the calc_lambda function.
The plotting functions for anisotropy and intensity graphs were integrated into the FA class. The plots of anisotropy and intensity can be plotted side by side on a grid for all repeats (plot_ani_all function). Plot of anisotropy or intensity for a single protein-tracer pair can be displayed as well (plot_ani function). The plots can be exported (export_ani_figs function). The export_ani_figs and plot_ani functions use the same _plot_ani function which plots just one figure.
Fixed the error in the SciPy curve_fit function occurring when there are NaNs in the data to be fitted (as a result of user invalidation) by removing all NaN values from the respective data frame before fitting.