Update fluores_anisotropy.ipynb
A warning was added to the logistic_fit_all and logistic_fit_pair functions to warn the user if the fitting failed for any of the protein-tracer pairs, a message providing the details of the failures is also displayed.
The button handler function now keeps track of protein-tracer pairs for which the lambda value was saved so that lambda value from only one repeat is saved. Also, a final_fit data frame was added as a class variable for storing the final values of rmin, rmax and lambda for each protein-tracer pair irrespectively of the repeat number. The button handler function now adds those variables to the final_fit data frame.
The mechanism for plotting the anisotropy and intensity was changed. Now there is only one function (_plot) that deals with plotting given variables, labelling the plot and saving it. The three other functions for displaying all plots on a grid, exporting all plots as single figures and displaying a single figure call the _plot function with different arguments.