Update fluores_anisotropy.ipynb
The functionality of the ‘calc_lambda’ function was extended. Another column of checkboxes with the rmin and rmax values was added next to the lambda checkboxes so that the user can check the rmin and rmin even if the lambda value is not chosen for a given protein-tracer pair. Only the first ticked lambda value is saved for a given protein tracer-pair. The functionality of generating the widget and handling widget events was moved into a separate _widget function.
The ‘export_params’ function was created so that the fitting parameters for each repeat can be exported into a csv or excel file.
The ‘import_params’ function was added so that the final fitting parameters (rmin, rmax, lambda) can be imported from a csv file.