Update fluores_anisotropy.ipynb
The final_fit data frame is now created during the instantiation of the class instead of during the calculation of the data for fitting so that the  final fitting parameters can be immediately imported.
The 'export_params' function now also exports the final fit_fit data frame alongside the fitting params data frames for each repeat.
In the 'import_params' function a warning is raised instead of the error.