Update fluores_anisotropy.ipynb
Different colour maps are used to display each data data set in the overlaid plot of binding constant. The legend was moved outside of the plotting area.
New function '_get_items_to_plot' was created to handle the protein and tracer names and repeat number arguments provided to the plotting functions.
The anisotropy plotting functions were modified to accept the same arguments as the binding constant plotting functions.
Handling of the p0 and sigma arguments was moved to the _logistic_fit function.