Update fluores_anisotropy.ipynb
The _background_correct function was integrated into the FA class, it was adjusted so that it works for both protein and tracer titrations and competitor experiments.
The _calc_mean_r_i and _calc_amount_bound functions were adjusted to work with competition data as well. The data frames are now named with a 3-tuple (protein, tracer and competitor) instead of just protein and tracer.
Another column with competitor name is now displayed on the widget with lambda values.
New function _r_func2 was created for curve fitting  the competitor data.
The logistic_fit function was adjusted so that it used the _r_func2 to do the fitting of the anisotropy against the competitor concertation.
The plot_ani function was adjusted to plot the anisotropy vs competitor concentration for the competitor experiments.
Final fit and fitting parameters data frames are now indexed by protein, tracer and competitor names.