Update fluores_anisotropy.ipynb
The button event function was modified so that now the third part of the index is ‘-‘ so that the params are assigned only by the specific protein-tracer name and cannot be assigned if there is a competitor (Index error after hitting the save button).
The _import_params function was modified so that it in case of the imported data frame does not contain a competitor name, the final_fit data frame is updated only y protein and tracer names.
The _calc_amount_bound function was updated. Now only the rmin, rmax and lambda columns are joined to the plate map with anisotropy values instead of the whole final_fit data frame.
A new function for curve fitting of the plot of fluorescent tracer bound against competitor concentration was created.
A new function for plotting the fluorescent tracer bound against competitor concentration was created.
A new function for calculating the Ki was created.