Update fluores_anisotropy.ipynb
The _get_items_to plot function was modified to include the competitor names. All functions using this function were modified to include the com parameter. The com argument can only be called in the functions specific to the competition data processing.

The _calc_Ki function was created for calculating the Ki and its error. It was implemented into the plot_ic50 function so that the Ki is calculated and plotted on the ic50 plots.