Update fluores_anisotropy.ipynb
Formatting of the Ki values displayed on the graph was changed to include the comma as a thousand separator to improve the readability.
The 'plot_ic50' function now iterates over the whole data dictionary and calculates the Ki for all repeats and protein-tracer pairs but the figures are only plotted for the competitors specified by the user.
The  '_calc_ic50' function now uses the Pt parameter to calculate the theoretical value of Ki instead of the p0 parameter.