Update
The ‘wellcolor’ function was modified, an option for generating a normalised array of colours was added.

In the ‘lebelwell’ function another option for formatting the percentage intensity values was added.

The ‘scinot’ parameter was replaced by ‘str_format’ in all functions. It can be either ‘scinot’ or ‘percent’ and regulates the formatting of a number.