Update fluores_anisotropy.ipynb
The warning in plot_ani_com and plot_ic50 functions was suppressed. Serval variables were renamed to improve readability. The read_in_enivision function now displays ‘None’ for the competitors instead of [‘-‘] for protein or tracer titration. The _r_func and _r_func2 functions were renamed into _EC50 and _EC50_com for better clarity. The annotation boxes were removed in all figures, all information is now displayed inside the legend boxes. The information about repeat number does not appear in the figure title or legends if there is only one repeat. Potentially large parameters displayed in the legend boxes are now formatted with a comma as a thousand separator in all plotting functions. The units displayed on all figures are now taken from the platemap.  An option to save the competition anisotropy figures was added to plot_ani_com function. The export_data function was updated to include the export of the competition data. The export_params function was updated to include the export of the competition fitting parametrs. A warning was implemented into set_fitparams function when the paraments passed as keyword arguments do not match any column in the final_fit data frame. A separate function vir_data was created to generate the array of concentration data points used for plotting of the fitted curve to avoid repeating the same code over multiple plotting functions. The nan is now replaced with ‘-‘ when generating the protein-tracer-competitor tuples in the _get_items_to_plot function to avoid errors in there no competitors in the assay. The fitting_params_com data frame is now created only if there are competitors in the platemap. In the visualise function the non-categorical variables and the ones potentially requiring sicnot formatting are defined separately at the beginning of the function.